refactor:update naming structure
